Make rate change work for play example
Fix problem with premature EOF by setting length on signal for output
effect in chain, Not sure I understand why that's the fix

Still have an extra free and had to do the sox_init to get past
the fft 0 problem.

Feels like one more clean up of the library and the signal handling will
be all right.